+++ src/cli/typescript-toolkit-cli.ts
@@ -2,6 +2,7 @@ import yargs from "yargs";
 
 import { addBuildCommands } from "./commands/build";
 import { addCleanCommands } from "./commands/clean";
+import { addFileCommands } from "./commands/file";
 import { addFormatCommands } from "./commands/format";
 import { addInstallCommands } from "./commands/install";
 import { addLintCommands } from "./commands/lint";
@@ -19,6 +20,7 @@ export class TypescriptToolkitCLI {
     addInstallCommands({ to: builder });
     addLintCommands({ to: builder });
     addTestCommands({ to: builder });
+    addFileCommands({ to: builder });
 
     builder.demandCommand().version();